add hive cluster task
