+++ linkis-computation-governance/linkis-engineconn/linkis-computation-engineconn/src/main/scala/org/apache/linkis/engineconn/computation/executor/rs/RsOutputStream.scala
@@ -25,6 +25,7 @@ import org.apache.linkis.engineconn.computation.executor.execute.EngineExecution
 import org.apache.linkis.storage.LineRecord
 
 import java.io.OutputStream
+
 import scala.collection.mutable.ArrayBuffer
 
 class RsOutputStream extends OutputStream with Logging {
