+++ linkis-computation-governance/linkis-manager/linkis-application-manager/src/main/scala/org/apache/linkis/manager/rm/service/impl/DriverAndYarnReqResourceService.scala
@@ -30,6 +30,10 @@ import org.apache.linkis.manager.rm.external.service.ExternalResourceService
 import org.apache.linkis.manager.rm.external.yarn.YarnResourceIdentifier
 import org.apache.linkis.manager.rm.service.{LabelResourceService, RequestResourceService}
 import org.apache.linkis.manager.rm.utils.{AcrossClusterRulesJudgeUtils, RMUtils}
+import org.apache.linkis.manager.rm.utils.AcrossClusterRulesJudgeUtils.{
+  originClusterResourceCheck,
+  targetClusterResourceCheck
+}
 
 import org.apache.commons.lang3.StringUtils
 
@@ -119,125 +123,15 @@ class DriverAndYarnReqResourceService(
     }
     if (engineCreateRequest.getProperties != null) {
       // judge if is cross cluster task and target cluster priority first
-      targetClusterResourceCheck(labelContainer, engineCreateRequest, maxCapacity, usedCapacity)
+      targetClusterResourceCheck(
+        labelContainer,
+        engineCreateRequest,
+        maxCapacity,
+        usedCapacity,
+        externalResourceService
+      )
     }
     true
-  }
-
-  def originClusterResourceCheck(
-      engineCreateRequest: EngineCreateRequest,
-      maxCapacity: Resource,
-      usedCapacity: Resource
-  ): Unit = {
-    val acrossClusterTask =
-      engineCreateRequest.getProperties.getOrDefault(AMConfiguration.ACROSS_CLUSTER_TASK, "false")
-    val priorityCluster = engineCreateRequest.getProperties.get(AMConfiguration.PRIORITY_CLUSTER)
-    if (
-        StringUtils.isNotBlank(acrossClusterTask) && acrossClusterTask.toBoolean && StringUtils
-          .isNotBlank(priorityCluster) && priorityCluster.equals(
-          AMConfiguration.PRIORITY_CLUSTER_ORIGIN
-        )
-    ) {
-      // get origin cluster resource threshold
-      val originCPUPercentageThreshold =
-        engineCreateRequest.getProperties.get(AMConfiguration.ORIGIN_CPU_PERCENTAGE_THRESHOLD)
-      val originMemoryPercentageThreshold =
-        engineCreateRequest.getProperties.get(AMConfiguration.ORIGIN_MEMORY_PERCENTAGE_THRESHOLD)
-      if (
-          StringUtils.isNotBlank(originCPUPercentageThreshold) && StringUtils.isNotBlank(
-            originMemoryPercentageThreshold
-          )
-      ) {
-        // judge origin cluster resource in origin threshold
-        try {
-          AcrossClusterRulesJudgeUtils.originClusterRuleCheck(
-            usedCapacity.asInstanceOf[YarnResource],
-            maxCapacity.asInstanceOf[YarnResource],
-            originCPUPercentageThreshold.toDouble,
-            originMemoryPercentageThreshold.toDouble
-          )
-        } catch {
-          // if origin cluster resource gt threshold, throw origin retry exception and change to target cluster next retry;
-          case ex: Exception =>
-            throw new RMWarnException(
-              RMErrorCode.ACROSS_CLUSTER_RULE_FAILED.getErrorCode,
-              ex.getMessage
-            )
-        }
-      }
-    }
-  }
-
-  def targetClusterResourceCheck(
-      labelContainer: RMLabelContainer,
-      engineCreateRequest: EngineCreateRequest,
-      maxCapacity: Resource,
-      usedCapacity: Resource
-  ): Unit = {
-    val acrossClusterTask =
-      engineCreateRequest.getProperties.getOrDefault(AMConfiguration.ACROSS_CLUSTER_TASK, "false")
-    val priorityCluster = engineCreateRequest.getProperties.get(AMConfiguration.PRIORITY_CLUSTER)
-    if (
-        StringUtils.isNotBlank(acrossClusterTask) && acrossClusterTask.toBoolean && StringUtils
-          .isNotBlank(priorityCluster) && priorityCluster.equals(
-          AMConfiguration.PRIORITY_CLUSTER_TARGET
-        )
-    ) {
-      val leftResource = maxCapacity - usedCapacity
-      // get target cluster resource threshold
-      val targetCPUThreshold =
-        engineCreateRequest.getProperties.get(AMConfiguration.TARGET_CPU_THRESHOLD)
-      val targetMemoryThreshold =
-        engineCreateRequest.getProperties.get(AMConfiguration.TARGET_MEMORY_THRESHOLD)
-      val targetCPUPercentageThreshold =
-        engineCreateRequest.getProperties.get(AMConfiguration.TARGET_CPU_PERCENTAGE_THRESHOLD)
-      val targetMemoryPercentageThreshold =
-        engineCreateRequest.getProperties.get(AMConfiguration.TARGET_MEMORY_PERCENTAGE_THRESHOLD)
-      val clusterCPUPercentageThreshold =
-        AMConfiguration.ACROSS_CLUSTER_TOTAL_CPU_PERCENTAGE_THRESHOLD
-      val clusterMemoryPercentageThreshold =
-        AMConfiguration.ACROSS_CLUSTER_TOTAL_MEMORY_PERCENTAGE_THRESHOLD
-      if (
-          StringUtils
-            .isNotBlank(targetCPUThreshold) && StringUtils.isNotBlank(targetMemoryThreshold)
-          && StringUtils.isNotBlank(targetCPUPercentageThreshold) && StringUtils.isNotBlank(
-            targetMemoryPercentageThreshold
-          )
-      ) {
-        // judge total target cluster resources in target threshold
-        val clusterYarnResource =
-          externalResourceService.getResource(
-            ResourceType.Yarn,
-            labelContainer,
-            new YarnResourceIdentifier("root")
-          )
-        val (clusterMaxCapacity, clusterUsedCapacity) =
-          (clusterYarnResource.getMaxResource, clusterYarnResource.getUsedResource)
-        // judge target cluster resource in target threshold
-        try {
-          AcrossClusterRulesJudgeUtils.targetClusterRuleCheck(
-            leftResource.asInstanceOf[YarnResource],
-            usedCapacity.asInstanceOf[YarnResource],
-            maxCapacity.asInstanceOf[YarnResource],
-            clusterMaxCapacity.asInstanceOf[YarnResource],
-            clusterUsedCapacity.asInstanceOf[YarnResource],
-            targetCPUThreshold.toInt,
-            targetMemoryThreshold.toInt,
-            targetCPUPercentageThreshold.toDouble,
-            targetMemoryPercentageThreshold.toDouble,
-            clusterCPUPercentageThreshold,
-            clusterMemoryPercentageThreshold
-          )
-        } catch {
-          // if target cluster resource gt threshold, throw target retry exception and change to normal task next retry;
-          case ex: Exception =>
-            throw new RMWarnException(
-              RMErrorCode.ACROSS_CLUSTER_RULE_FAILED.getErrorCode,
-              ex.getMessage
-            )
-        }
-      }
-    }
   }
 
   def generateQueueNotEnoughMessage(
